chore: Move HF EF to pkg/embeddings
