+++ pkg/embeddings/hf/option.go
