+++ test/chroma_client_test.go
@@ -11,6 +11,7 @@ import (
 	"context"
 	"encoding/base64"
 	"fmt"
+	"github.com/amikos-tech/chroma-go/pkg/embeddings/hf"
 	"github.com/testcontainers/testcontainers-go"
 	tcchroma "github.com/testcontainers/testcontainers-go/modules/chroma"
 	"net/http"
@@ -28,7 +29,6 @@ import (
 	chroma "github.com/amikos-tech/chroma-go"
 	"github.com/amikos-tech/chroma-go/cohere"
 	"github.com/amikos-tech/chroma-go/collection"
-	"github.com/amikos-tech/chroma-go/hf"
 	"github.com/amikos-tech/chroma-go/types"
 	"github.com/amikos-tech/chroma-go/where"
 	wheredoc "github.com/amikos-tech/chroma-go/where_document"
